+++ R/3_1_talkTranscribeDiarise.R
@@ -165,3 +165,7 @@ talkTranscribeDiarise <- function(
     show = TRUE
   )
 }
+
+#' @rdname talkTranscribeDiarise
+#' @export
+talkTextDiarise <- talkTranscribeDiarise